SREP-1037: Add Tests, and fix namespace controller defect
Adds tests for SREP-1037 and adds a fix in the namespace controller to
allow for the case when the configuration ConfigMap is absent, or does
not otherwise have the expected `project_blacklist` key. The fix is to
requeue the request after 5 seconds.

This change also modifies the Makefile to add a `test` target, and to
force `build` to depend upon `test`.

Signed-off-by: Lisa Seelye <lseelye@redhat.com>