SREP-1039: Add Tests, and fix namespace controller defect (#13)
* SREP-1039: Add Tests, update vendor deps

This is going to be a bit of an odd merge, given the outstanding comments both here and @jewzaam's proposed PR in https://github.com/lisa/dedicated-admin-operator/pull/1 but in light of #16 it makes sense.

As noted in https://github.com/lisa/dedicated-admin-operator/pull/1#discussion_r270998409 we have a potential issue where an invalid regular expression is specified in the configuration and we then have weird behaviour when trying to match against it (and how do we handle it?), and dealing with what we do when users edit the `ConfigMap` at runtime? 

With #16, however, we're going to move away from a `ConfigMap` entirely, in favour of a hardcoded set of compiled regular expressions in the operator's source. Thus, if the compilation fails the operator won't start, and we won't have unexpected behaviour. Plus, this moves some guarantees to `make test` and `make build`) to ensure the provided (static) config is safe.

This PR is being merged because of the large number of files changed (thanks `vendor` 😦 ) and how painful a rebase after #16 is completed. We know that these tests will have to change for #16 and that's okay.


Signed-off-by: Lisa Seelye <lseelye@redhat.com>