Fix ClusterRoleBindings for OLM
Figured out that some of the ClusterRoleBindings in this repo didn't have kind in the roleRef.  OLM doesn't like this, hence inability to deploy them from the bundle. I've fixed it an tested with osd-operators-registry and will put in a PR to check that they're well formed when generating the catalog stuff.