Rework what is bundled in catalog vs in syncset
OLM can deploy a resource that is bundled in a CatalogSource.  But OLM won't update it.  And OLM won't delete it if it's removed in a future version of the CatalogSource. There are two other options for managing these resources.  1) controller in the operator code or 2) manage it externally.  Option #1 is a lot of work, though it does cover the cases where a resource is deleted.  Option #2 is less work but has a gap in that fixing a broken config requires external action.

This PR is going with option #2 and allowing Hive to manage the resources via SelectorSyncSet.  Hive will fix anything that breaks within 2 hours or a human can force it to sync by removing the related SyncSetInstance CR.  It means no go code to manage the resources and a simpler deployment.  This means ALL resources move out of the bundle.

We can go back to bundling in the future when OLM will manage bundled resources.  It's causing pain now and I have been fixing things by creating one-off SelectorSyncSets to remove stale resources in an environment.  This will mean no longer doing this, Hive does that work for us.