:bug: Compare version to correct thing
On line 36, we get a value looking something like this `0.1.166-1e758c1` and then immediately check if it exists, if it does, we reset it to `$(OPERATOR_NAME).$$version`, in this case `dedicated-admin-operator.v0.1.166-1e758c1`. Then, we build the new bundle and on line 43 detect what that bundle was.

Line 43 gives us something looking like `0.1.166-1e758c1`. On line 44 we try to check if this is a re-build, and if it is, we should bail out. But, we're comparing `dedicated-admin-operator.v0.1.166-1e758c1 to 0.1.166-1e758c1`.

This change, on line 44, makes sure to compare like to like.